+++ R/gaussianSamples.R
@@ -8,7 +8,9 @@
 #' @param pi0 Proportion of true null hypotheses
 #' @param SNR Signal to noise ratio. Either a numeric value (a measure of
 #'   distance between H0 and H1) or a vector of length \code{m*(1-pi0)}
-#' @param p A numeric value in (0, 1], the frequency of the population under H1. If \eqn{p=1} (the default), a data set from a single population is generated.
+#' @param prob A numeric value in (0, 1], the frequency of the population 
+#'   under H1. If \eqn{prob=1} (the default), a data set from a single 
+#'   population is generated.
 #' @param w An optional vector of length \code{n}, the underlying factor driving
 #'   equi-correlation
 #' @return A list with two elements: \describe{
@@ -17,7 +19,7 @@
 #' hypothesis: 0 for true null hypothesis, and 1 for true alternative
 #' hypothesis}}.
 #' @author Gilles Blanchard, Pierre Neuvial and Etienne Roquain
-#' @details. If \code{p = 1}, each of m1 variables under H1 has mean \eqn{SNR/\sqrt(n)}. If \code{0 < p < 1} n0 samples are drawn from a N(0,1) distribution while n1 are drawn from a N(mu, 1) distribution, where \eqn{mu = SNR*sqrt(1/(n0) + 1/n1)}. The argument \code{p} is the probability of a sample to belong to the non-zero-mean population
+#' @details. If \code{prob = 1}, each of m1 variables under H1 has mean \eqn{SNR/\sqrt(n)}. If \code{0 < p < 1} n0 samples are drawn from a N(0,1) distribution while n1 are drawn from a N(mu, 1) distribution, where \eqn{mu = SNR*sqrt(1/(n0) + 1/n1)}. The argument \code{p} is the probability of a sample to belong to the non-zero-mean population
 #' @export
 #' @importFrom stats rbinom
 #' @examples
@@ -29,8 +31,8 @@
 #' B <- 1e3
 #'
 #' ## two-sample data
-#' sim <- gaussianSamples(m, rho, n, pi0, SNR=2, p=0.5)
-#' tests <- testByRandomization(sim$X, B, flavor = "perm")
+#' sim <- gaussianSamples(m, rho, n, pi0, SNR = 2, prob = 0.5)
+#' tests <- testByRandomization(sim$X, B)
 #'
 #' ## show test statistics
 #' pch <- 20
@@ -38,9 +40,8 @@
 #' plot(tests$T, col=colStat, main="Test statistics", pch=pch)
 #' legend("topleft", c("H0", "H1"), pch=pch, col=1:2)
 #'
-#' ## sign-flipping
 #' sim <- gaussianSamples(m, rho, n, pi0, SNR=2)
-#' tests <- testByRandomization(sim$X, B, flavor = "flip")
+#' tests <- testByRandomization(sim$X, B)
 #'
 #' ## show test statistics
 #' pch <- 20
@@ -48,7 +49,7 @@
 #' plot(tests$T, col=colStat, main="Test statistics", pch=pch)
 #' legend("topleft", c("H0", "H1"), pch=pch, col=1:2)
 #'
-gaussianSamples <- function(m, rho, n, pi0, SNR = 1, p = 1, w = NULL) {
+gaussianSamples <- function(m, rho, n, pi0, SNR = 1, prob = 1, w = NULL) {
     m0 <- round(m*pi0)
     m1 <- m - m0
     H <- rep(c(0, 1), times = c(m0, m1))
@@ -56,7 +57,7 @@ gaussianSamples <- function(m, rho, n, pi0, SNR = 1, p = 1, w = NULL) {
 
     ## sanity checks
     if (length(SNR) > 1) {
-        stopifnot(length(SNR)==m1)
+        stopifnot(length(SNR) == m1)
     }
 
     ## 1. equi-correlated noise
@@ -66,13 +67,13 @@ gaussianSamples <- function(m, rho, n, pi0, SNR = 1, p = 1, w = NULL) {
     ## 2. signal
     mu <- matrix(0, nrow = nrow(eps), ncol = ncol(eps)) ## m x n
     if (m0 < m) {
-        if (p == 1) {  
+        if (prob == 1) {  
             # one-sample data (defined here as a corner case of the two-sample problem)
             signal <- SNR/sqrt(n)
             mu[H1, ] <- signal
             colnames(mu) <- rep(1, n)
         } else {
-            y <- rbinom(n, 1, p)     ## Bernoulli response
+            y <- rbinom(n, 1, prob)     ## Bernoulli response
             w1 <- which(y == 1)
             n1 <- length(w1)
             signal <- SNR*sqrt(1/(n-n1) + 1/n1)  ## ! scaling is test-specific (here T)!