+++ R/testByRandomization.R
@@ -1,14 +1,10 @@
 #' Randomization-based testing
-#' 
-#' Randomization-based testing using permutation or sign-flipping 
+#'
+#' Randomization-based testing using permutation or sign-flipping
 #'
 #' @param X a matrix of \code{m} variables by \code{n} observations
 #'
 #' @param B A numeric value, the number of permutations to be performed
-#'
-#' @param flavor A character value, the type of randomization to be performed.
-#'   Should either be "perm" for two-sample permutation or "flip" for sign
-#'   flipping.
 #'
 #' @param cls A vector of length \code{n} class labels in \code{0,1} for flavor
 #'   "perm". Defaults to colnames(X).
@@ -19,23 +15,27 @@
 #' @param seed An integer (or NULL) value used as a seed for random number
 #'   generation. If \code{NULL}, no seed is specified
 #'
-#' @details If \code{flavor == "permutation"}, we test the null hypothesis: "both groups have the
-#' same mean" against the one-sided alternative that the mean is larger in the
-#' second group. The test is Welch's two-sample
-#'   test for unequal variances. Permuted test statistics
-#' are calculated by B permutations of the group labels. Corresponding observed
-#' and permuted p-values are calculated as the proportion of permutations
-#' (including the identity) for which the permuted test statistic is larger than
-#' the observed test statistic. The corresponding parametric p-values are
-#'   returned as elements \code{param.p} (for the original data) and
-#'   \code{param.p0} (for the permutations).
-#' 
-#' @details If \code{flavor == "flip"}, we test the null hypothesis: "the mean is 0" against the
-#' two-sided alternative that the mean is larger than 0. We use the (rescaled)
-#' empirical mean of the observations as a test statistic. Sign-flipped test
-#' statistics are calculated by flipping the sign of each observation with
-#' probability 1/2.
+#' @details The type of randomization is determined by the parameter \code{cls}.
+#'   If \code{cls} does not containt two distinct values (or is \code{NULL}), a
+#'   one-sample test is performed using randomization (flavor "flip"). If it
+#'   contains two distinct values, a two-sample test is perfomed using
+#'   permutations (flavor "perm").
 #'
+#'   For permutation, we test the null hypothesis: "both groups have the same
+#'   mean" against the one-sided alternative that the mean is larger in the
+#'   second group. The test is Welch's two-sample test for unequal variances.
+#'   Permuted test statistics are calculated by B permutations of the group
+#'   labels. Corresponding observed and permuted p-values are calculated as the
+#'   proportion of permutations (including the identity) for which the permuted
+#'   test statistic is larger than the observed test statistic. The
+#'   corresponding parametric p-values are returned as elements \code{param.p}
+#'   (for the original data) and \code{param.p0} (for the permutations).
+#'
+#'   For sign-flipping, we test the null hypothesis: "the mean is 0" against the
+#'   two-sided alternative that the mean is larger than 0. We use the (rescaled)
+#'   empirical mean of the observations as a test statistic. Sign-flipped test
+#'   statistics are calculated by flipping the sign of each observation with
+#'   probability 1/2.
 #'
 #' @references Ge, Y., Dudoit, S. and Speed, T.P., 2003. Resampling-based
 #'   multiple testing for microarray data analysis. _Test_, 12(1), pp.1-77.
@@ -47,20 +47,27 @@
 #'   \item{T0}{A \eqn{m \times B} matrix of randomized test statistics}
 #'
 #'   \item{p}{A vector of \eqn{m} \eqn{p}-values (only if \code{p.value} is
-#' \code{TRUE} )}
+#'   \code{TRUE} )}
 #'
 #'   \item{p0}{A \eqn{m \times B} matrix of randomization \eqn{p}-values (only
 #'   if \code{p.value} is \code{TRUE} )}
 #'
-#'   \item{param.p}{A vector of \eqn{m} parametric \eqn{p}-values (only for flavor
-#'   "permutation")}
+#'   \item{flavor}{A character value, the type of randomization performed:
+#'   "perm" for permutation-based randomization in two-sample tests, and "flip"
+#'   for sign-flipping-based randomization in one sample tests. See Details.}
 #'
-#'   \item{param.p0}{A \eqn{m \times B} matrix of parametric \eqn{p}-values on permuted data (only for flavor "permutation" )} 
-#'   
-#'   \item{df}{A vector of \eqn{m} degrees of freedom for the observed statistics (only for flavor "permutation")}
+#'   \item{param.p}{A vector of \eqn{m} parametric \eqn{p}-values (only for
+#'   flavor "perm")}
 #'
-#'   \item{df0}{A \eqn{m \times B} matrix of degrees of freedom on permuted data (only
-#'   for flavor "permutation" )} }
+#'   \item{param.p0}{A \eqn{m \times B} matrix of parametric \eqn{p}-values on
+#'   permuted data (only for flavor "perm" )}
+#'
+#'   \item{df}{A vector of \eqn{m} degrees of freedom for the observed
+#'   statistics (only for flavor "perm")}
+#'
+#'   \item{df0}{A \eqn{m \times B} matrix of degrees of freedom on permuted data
+#'   (only for flavor "perm" )}}
+#'
 #'
 #' @examples
 #'
@@ -69,8 +76,8 @@
 #' B <- 1e3
 #' mat <- matrix(rnorm(p*n), ncol=n)
 #' cls <- rep(c(0, 1), times=c(n/2, n-n/2))
-#' resPerm <- testByRandomization(X=mat, flavor="perm", cls=cls, B=B)
-#' resFlip <- testByRandomization(X=mat, flavor="flip", B=B)
+#' resPerm <- testByRandomization(X=mat, cls=cls, B=B) ## permutation
+#' resFlip <- testByRandomization(X=mat, B=B)          ## sign-flipping
 #'
 #' # empirical coverage of Simes thresholds
 #' alpha <- 0.2
@@ -78,10 +85,10 @@
 #' sansSouci:::empiricalCoverage(thr, resPerm$T0) ## Welch, not Gaussian
 #' sansSouci:::empiricalCoverage(thr, qnorm(1-resPerm$p0))
 #' sansSouci:::empiricalCoverage(thr, qnorm(1-resPerm$param.p0))
-#' 
+#'
 #' sansSouci:::empiricalCoverage(thr, resFlip$T0)
 #' sansSouci:::empiricalCoverage(thr, qnorm(1-resFlip$p0))
-#' 
+#'
 #' # test statistics null distribution
 #' hist(resPerm$p)
 #' hist(resFlip$p)
@@ -93,18 +100,29 @@
 #' @importFrom matrixStats rowRanks
 #' @export
 #' 
-testByRandomization <- function(X, B, flavor = c("perm", "flip"), 
-                                cls = colnames(X), 
+testByRandomization <- function(X, B, cls = colnames(X), 
                                 p.value = TRUE, seed = NULL){
     ## sanity checks
     n <- ncol(X)
-    flavor <- match.arg(flavor)
-    if (flavor == "perm") {
+    luc <- length(unique(cls))
+    if (luc <= 1) {  
+        # no classes or a single class given: assuming sign flipping 
+        flavor <- "flip"
+    } else {
         if (length(cls) != n) {
             stop("The number of columns of argument 'X' should match the length of argument 'cls'")
         }
-        if (!all(sort(unique(cls)) == c(0, 1))) {
-            stop("Argument 'cls' should contain (only) 0:s and 1:s")
+        if (luc == 2) {
+            flavor <- "perm"
+            tbl <- table(cls)
+            if ( !all(names(tbl) == c("0", "1"))) {  # note that numeric values are allowed in cls as they are converted into character by 'table'...
+                stop("Argument 'cls' should contain (only) 0:s and 1:s")
+            }
+            if (min(tbl) < 3) {
+                stop("Argument 'cls' should contain at least 3 elements of each sample")
+            }
+        } else if (luc > 2) {
+            stop("Tests for more than 2 classes not implemented yet")
         }
     }
     if (!is.null(seed)) {
@@ -134,6 +152,7 @@ testByRandomization <- function(X, B, flavor = c("perm", "flip"),
             df[, bb] <- rwt$parameter
         }
         res <- list(T = T_obs, T0 = T, 
+                    flavor = flavor,
                     param.p = p_obs, param.p0 = pp,
                     df = df_obs, df0 = df)
     } else if (flavor == "flip") {
@@ -141,9 +160,9 @@ testByRandomization <- function(X, B, flavor = c("perm", "flip"),
         T_obs <- rowSums(X)/sqrt(n)
         ## test statistics under H0
         T <- testBySignFlipping(X, B)
-        res <- list(T = T_obs, T0 = T)
+        res <- list(T = T_obs, T0 = T, flavor = flavor)
     }
-    
+
     if (p.value) {
         ## get m x (B+1) matrix of pvalues under the null (+ original)
         ## by sorting null test statistics as proposed by Ge et al (2003)